Move translator's ConnectionToContext to network
To reuse it later in the API HTTP server.
